+++ lib/solidpod.dart
@@ -43,9 +43,9 @@ export 'src/solid/popup_login.dart' show SolidPopupLogin;
 // following are exported, PLEASE. ReadPod() I understand, but what from
 // rest_api?
 
+export 'src/solid/api/rest_api.dart' show getFileContent;
+export 'src/solid/utils/misc.dart';
 export 'src/solid/change_key.dart' show changeKeyPopup;
 export 'src/widgets/logout_dialog.dart' show logoutPopup;
-export 'src/solid/api/rest_api.dart';
-export 'src/solid/utils.dart';
 export 'src/solid/read_pod.dart' show readPod;
 export 'src/solid/write_pod.dart' show writePod;