+++ lib/src/screens/initial_setup/initial_setup_screen_body.dart
@@ -40,7 +40,8 @@ import 'package:solid_auth/solid_auth.dart';
 import 'package:solidpod/src/screens/initial_setup/widgets/res_create_form_submission.dart';
 import 'package:solidpod/src/screens/initial_setup/widgets/enc_key_input_form.dart';
 import 'package:solidpod/src/screens/initial_setup/widgets/initial_setup_welcome.dart';
-import 'package:solidpod/src/solid/utils.dart' show AuthDataManager;
+import 'package:solidpod/src/solid/utils/authdata_manager.dart'
+    show AuthDataManager;
 
 /// A [StatefulWidget] that represents the initial setup screen for the desktop version of an application.
 ///