+++ lib/src/screens/initial_setup/widgets/res_create_form_submission.dart
@@ -40,12 +40,14 @@ import 'package:fast_rsa/fast_rsa.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:solid_encrypt/solid_encrypt.dart' as solid_encrypt;
 
-import 'package:solidpod/src/screens/initial_setup/initial_setup_constants.dart';
 import 'package:solidpod/src/screens/initial_setup/gen_file_body.dart';
+import 'package:solidpod/src/screens/initial_setup/initial_setup_constants.dart';
 import 'package:solidpod/src/solid/api/rest_api.dart';
-import 'package:solidpod/src/solid/utils.dart'
-    show AuthDataManager, getWebId, saveMasterPassword;
+import 'package:solidpod/src/solid/utils/authdata_manager.dart'
+    show AuthDataManager;
 import 'package:solidpod/src/solid/constants.dart';
+import 'package:solidpod/src/solid/utils/misc.dart'
+    show getWebId, saveMasterPassword;
 import 'package:solidpod/src/widgets/error_dialog.dart';
 import 'package:solidpod/src/widgets/show_animation_dialog.dart';