+++ lib/src/solid/utils/misc.dart
@@ -1,4 +1,4 @@
-/// Common utility functions used across the package.
+/// Miscellaneous utility functions used across the package.
 ///
 // Time-stamp: <Friday 2024-02-16 10:59:10 +1100 Graham Williams>
 ///
@@ -32,21 +32,18 @@ library;
 
 import 'dart:convert';
 
-import 'package:flutter/foundation.dart' hide Key;
-
 import 'package:crypto/crypto.dart';
 import 'package:encrypt/encrypt.dart';
-import 'package:fast_rsa/fast_rsa.dart';
 import 'package:jwt_decoder/jwt_decoder.dart';
-import 'package:package_info_plus/package_info_plus.dart';
 import 'package:path/path.dart' as path;
 import 'package:rdflib/rdflib.dart';
 import 'package:solid_auth/solid_auth.dart';
-import 'package:solid_auth/src/openid/openid_client.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 import 'package:solidpod/src/solid/api/rest_api.dart';
 import 'package:solidpod/src/solid/constants.dart';
+import 'package:solidpod/src/solid/utils/app_info.dart' show AppInfo;
+import 'package:solidpod/src/solid/utils/authdata_manager.dart'
+    show AuthDataManager;
 
 // solid-encrypt uses unencrypted local storage and refers to http://yarrabah.net/ for predicates definition,
 // do not use it before it is updated (same as what the gurriny project does)
@@ -93,7 +90,7 @@ Future<String?> getVerificationKey() async {
     return null;
   }
 
-  final encKeyFileUrl = await getResourceUrl(encKeyPath);
+  final encKeyFileUrl = await getFileUrl(encKeyPath);
   if (!encKeyMap.containsKey(encKeyFileUrl)) {
     return null;
   }
@@ -107,20 +104,12 @@ bool verifyMasterPassword(String masterPasswd, String verificationKey) =>
     genVerificationKey(masterPasswd) == verificationKey;
 
 /// Save master password to local secure storage
-Future<void> saveMasterPassword(String masterPasswd) async {
-  await writeToSecureStorage(masterPasswdSecureStorageKey, masterPasswd);
-}
+Future<void> saveMasterPassword(String masterPasswd) async =>
+    writeToSecureStorage(masterPasswdSecureStorageKey, masterPasswd);
 
 /// Load master password from local secure storage
-Future<String?> loadMasterPassword() async {
-  // final webId = await getWebId();
-  // assert(webId != null);
-  // TODO: the current initialisation code uses web ID as key, update it.
-  // see src/screens/initial_setup/widgets/res_create_form_submission.dart
-  final masterPasswd =
-      await secureStorage.read(key: masterPasswdSecureStorageKey);
-  return masterPasswd;
-}
+Future<String?> loadMasterPassword() async =>
+    secureStorage.read(key: masterPasswdSecureStorageKey);
 
 /// Delete the saved master password from local secure storage
 Future<void> removeMasterPassword() async {
@@ -130,11 +119,8 @@ Future<void> removeMasterPassword() async {
 }
 
 /// Encrypt data using AES with the specified key
-String encryptData(String data, Key key, IV iv) {
-  final encrypter = Encrypter(AES(key));
-  final encryptVal = encrypter.encrypt(data, iv: iv);
-  return encryptVal.base64;
-}
+String encryptData(String data, Key key, IV iv) =>
+    Encrypter(AES(key)).encrypt(data, iv: iv).base64;
 
 /// Decrypt a ciphertext value
 String decryptData(String encData, Key key, IV iv) =>
@@ -168,7 +154,7 @@ Map<String, dynamic> parseTTL(String ttlContent) {
 
 /// Load and parse a private TTL file from POD
 Future<Map<String, dynamic>?> loadPrvTTL(String filePath) async {
-  final fileUrl = await getResourceUrl(filePath);
+  final fileUrl = await getFileUrl(filePath);
   try {
     final rawContent = await fetchPrvFile(fileUrl);
     return parseTTL(rawContent);
@@ -181,6 +167,8 @@ Future<Map<String, dynamic>?> loadPrvTTL(String filePath) async {
 /// Create a directory
 Future<bool> createDir(String dirName, String dirParentPath) async {
   try {
+    // await createItem(dirName,
+    //     itemLoc: dirParentPath, contentType: dirContentType, fileFlag: false);
     await createItem(false, dirName, '', fileLoc: dirParentPath);
     return true;
   } on Exception catch (e) {
@@ -195,6 +183,7 @@ Future<bool> createFile(String filePath, String fileContent) async {
     final fileName = path.basename(filePath);
     final folderPath = path.dirname(filePath);
 
+    // await createItem(fileName, itemLoc: folderPath, itemBody: fileContent);
     await createItem(true, fileName, fileContent,
         fileType: 'text/turtle', fileLoc: folderPath);
 
@@ -205,28 +194,29 @@ Future<bool> createFile(String filePath, String fileContent) async {
   return false;
 }
 
-/// Get the app name from pubspec.yml and
-/// 1. Remove any leading and trailing whitespace
-/// 2. Convert to lower case
-/// 3. Replace (one or multiple) white spaces with an underscore
-
-Future<String> getAppName() async {
-  final info = await PackageInfo.fromPlatform();
-  final appName = info.appName;
-  return appName.trim().toLowerCase().replaceAll(RegExp(r'\s+'), '_');
-}
-
-/// From a given resource path create its URL
-///
+/// From a given resource path [resourcePath] create its URL
+/// [isContainer] should be true if the resource is a directory, otherwise false
 /// returns the full resource URL
 
-Future<String> getResourceUrl(String resourcePath) async {
+Future<String> _getResourceUrl(String resourcePath, bool isContainer) async {
   final webId = await getWebId();
   assert(webId != null);
   assert(webId!.contains(profCard));
   final resourceUrl = webId!.replaceAll(profCard, resourcePath);
+  if (isContainer && !resourceUrl.endsWith('/')) {
+    return '$resourceUrl/';
+  }
+
   return resourceUrl;
 }
+
+/// Create the URL for a file
+Future<String> getFileUrl(String filePath) async =>
+    _getResourceUrl(filePath, false);
+
+/// Create the URL for a directory (container)
+Future<String> getDirUrl(String dirPath) async =>
+    _getResourceUrl(dirPath, true);
 
 /// Encrypt a given data string and format to TTL
 Future<String> getEncTTLStr(
@@ -234,8 +224,7 @@ Future<String> getEncTTLStr(
   final encData = encryptData(fileContent, key, iv);
 
   final g = Graph();
-  //final f = URIRef(appsFile + filePath); //TODO: update this
-  final f = URIRef(await getResourceUrl(filePath));
+  final f = URIRef(await getFileUrl(filePath));
   final ns = Namespace(ns: appsTerms);
   g.addTripleToGroups(f, ns.withAttr(pathPred), filePath);
   g.addTripleToGroups(f, ns.withAttr(ivPred), iv.base64);
@@ -258,22 +247,16 @@ Future<String> getEncTTLStr(
 }
 
 /// Returns the path of file with verification key and private key
-Future<String> getEncKeyPath() async {
-  final appName = await getAppName();
-  return path.join(appName, encDir, encKeyFile);
-}
+Future<String> getEncKeyPath() async =>
+    [await AppInfo.canonicalName, encDir, encKeyFile].join('/');
 
 /// Returns the path of file with individual keys
-Future<String> getIndKeyPath() async {
-  final appName = await getAppName();
-  return path.join(appName, encDir, indKeyFile);
-}
+Future<String> getIndKeyPath() async =>
+    [await AppInfo.canonicalName, encDir, indKeyFile].join('/');
 
 /// Returns the path of the data directory
-Future<String> getDataDirPath() async {
-  final appName = await getAppName();
-  return path.join(appName, dataDir);
-}
+Future<String> getDataDirPath() async =>
+    [await AppInfo.canonicalName, dataDir].join('/');
 
 /// Add (encrypted) individual/session key [encIndKey] and the corresponding
 /// IV [iv] for file with path [filePath]
@@ -282,7 +265,7 @@ Future<void> addIndKey(String filePath, String encIndKey, IV iv) async {
   // const termPrefix = '$appTermPrefix: <$appsTerms>';
   // final sub = appsFile + filePath;
   // final sub = '$appFilePrefix:$filePath';
-  final sub = await getResourceUrl(filePath);
+  final sub = await getFileUrl(filePath);
   // final query = [
   //   'PREFIX $filePrefix',
   //   'PREFIX $termPrefix',
@@ -295,17 +278,15 @@ Future<void> addIndKey(String filePath, String encIndKey, IV iv) async {
   //].join(' ');
   final query =
       'INSERT DATA {<$sub> <$appsTerms$pathPred> "$filePath"; <$appsTerms$ivPred> "${iv.base64}"; <$appsTerms$sessionKeyPred> "$encIndKey".};';
-  final fileUrl = await getResourceUrl(await getIndKeyPath());
+  final fileUrl = await getFileUrl(await getIndKeyPath());
   await updateFileByQuery(fileUrl, query);
 }
 
 /// Extract the app name and the version from the package info
 /// Return a record (with named fields https://dart.dev/language/records)
 
-Future<({String name, String version})> getAppNameVersion() async {
-  final info = await PackageInfo.fromPlatform();
-  return (name: info.appName, version: info.version);
-}
+Future<({String name, String version})> getAppNameVersion() async =>
+    (name: await AppInfo.name, version: await AppInfo.version);
 
 /// Check whether a user is logged in or not
 ///
@@ -334,213 +315,22 @@ Future<bool> checkLoggedIn() async {
 Future<bool> deleteLogIn() async {
   const success = true;
   try {
-    await secureStorage.delete(key: 'webid');
+    await secureStorage.delete(key: webIdSecureStorageKey);
   } on Exception {
     return false;
   }
   return success && (await AuthDataManager.removeAuthData());
 }
 
-/// Get the webId from local storage
+/// Save the webId to local secure storage
 
-Future<String?> getWebId() async {
-  final webId = await secureStorage.read(key: 'webid');
-  return webId;
-}
+Future<void> saveWebId(String webId) async =>
+    writeToSecureStorage(webIdSecureStorageKey, webId);
 
-/// [AuthDataManager] is a class to manage auth data returned by
-/// solid-auth authenticate, including:
-/// - save auth data to secure storage
-/// - load auth data from secure storage
-/// - delete saved auth data from secure storage
-/// - refresh access token if necessary
+/// Get the webId from local secure storage
 
-class AuthDataManager {
-  /// The URL for logging out
-  static String? _logoutUrl;
-
-  /// The RSA keypair and their JWK format
-  /// It seems Map<String, dynamic> does not work
-  static Map<dynamic, dynamic>? _rsaInfo;
-
-  /// The authentication response
-  static Credential? _authResponse;
-
-  /// The string key for storing auth data in secure storage
-  static const String _authDataSecureStorageKey = '_solid_auth_data';
-
-  /// Save the auth data returned by solid-auth authenticate in secure storage
-  /// It seems Map<String, dynamic> does not work
-  static Future<void> saveAuthData(Map<dynamic, dynamic> authData) async {
-    const keys = [
-      'client',
-      'rsaInfo',
-      'authResponse',
-      'tokenResponse',
-      'accessToken',
-      'idToken',
-      'refreshToken',
-      'expiresIn',
-      'logoutUrl',
-    ];
-
-    for (final key in keys) {
-      assert(authData.containsKey(key));
-    }
-
-    _logoutUrl = authData['logoutUrl'] as String;
-    _rsaInfo = authData['rsaInfo'] as Map<dynamic,
-        dynamic>; // Note that use Map<String, dynamic> does not seem to work
-    _authResponse = authData['authResponse'] as Credential;
-
-    await writeToSecureStorage(
-        _authDataSecureStorageKey,
-        jsonEncode({
-          'logout_url': _logoutUrl,
-          'rsa_info': jsonEncode({
-            ..._rsaInfo!,
-            // Overwrite the 'rsa' keypair in rsaInfo
-            'rsa': {
-              'public_key': _rsaInfo!['rsa'].publicKey as String,
-              'private_key': _rsaInfo!['rsa'].privateKey as String,
-            },
-          }),
-          'auth_response': _authResponse!.toJson(),
-        }));
-
-    debugPrint('AuthDataManager => saveAuthData() done');
-  }
-
-  /// Retrieve (and reconstruct) auth data from secure storage
-  /// It seems Map<String, dynamic> does not work
-  static Future<Map<dynamic, dynamic>?> loadAuthData() async {
-    if (_logoutUrl == null || _rsaInfo == null || _authResponse == null) {
-      final loaded = await _loadData();
-      if (!loaded) {
-        debugPrint('AuthDataManager => loadAuthData() failed');
-        return null;
-      }
-    }
-
-    assert(_logoutUrl != null && _rsaInfo != null && _authResponse != null);
-    try {
-      final tokenResponse = await _getTokenResponse();
-      return {
-        'client': _authResponse!.client,
-        'rsaInfo': _rsaInfo,
-        'authResponse': _authResponse,
-        'tokenResponse': tokenResponse,
-        'accessToken': tokenResponse!.accessToken,
-        'idToken': _authResponse!.idToken,
-        'refreshToken': _authResponse!.refreshToken,
-        'expiresIn': tokenResponse.expiresIn,
-        'logoutUrl': _logoutUrl,
-      };
-    } on Exception catch (e) {
-      debugPrint('AuthDataManager => loadAuthData() failed: $e');
-    }
-    return null;
-  }
-
-  /// Remove/delete auth data from secure storage
-  static Future<bool> removeAuthData() async {
-    try {
-      if (await secureStorage.containsKey(key: _authDataSecureStorageKey)) {
-        await secureStorage.delete(key: _authDataSecureStorageKey);
-        _logoutUrl = null;
-        _rsaInfo = null;
-        _authResponse = null;
-      }
-
-      return true;
-    } on Exception {
-      debugPrint('AuthDataManager => removeAuthData() failed');
-    }
-    return false;
-  }
-
-  /// Returns the (refreshed) access token
-  static Future<String?> getAccessToken() async {
-    final tokenResponse = await _getTokenResponse();
-    if (tokenResponse != null) {
-      return tokenResponse.accessToken;
-    } else {
-      debugPrint('AuthDataManager => getAccessToken() failed');
-    }
-    return null;
-  }
-
-  /// Returns the (updated) token response
-  static Future<TokenResponse?> _getTokenResponse() async {
-    if (_authResponse == null) {
-      final loaded = await _loadData();
-      if (!loaded) {
-        debugPrint('AuthDataManager => _getTokenResponse() failed');
-        return null;
-      }
-    }
-    assert(_authResponse != null);
-
-    try {
-      var tokenResponse = TokenResponse.fromJson(_authResponse!.response!);
-      if (JwtDecoder.isExpired(tokenResponse.accessToken!)) {
-        debugPrint(
-            'AuthDataManager => _getTokenResponse() refreshing expired token');
-        assert(_rsaInfo != null);
-        final rsaKeyPair = _rsaInfo!['rsa'] as KeyPair;
-        final publicKeyJwk = _rsaInfo!['pubKeyJwk'];
-        final tokenEndpoint =
-            _authResponse!.client.issuer.metadata['token_endpoint'] as String;
-        final dPopToken =
-            genDpopToken(tokenEndpoint, rsaKeyPair, publicKeyJwk, 'POST');
-        tokenResponse = await _authResponse!
-            .getTokenResponse(forceRefresh: true, dPoPToken: dPopToken);
-      }
-      return tokenResponse;
-    } on Exception catch (e) {
-      debugPrint('AuthDataManager => _getTokenResponse() failed: $e');
-    }
-    return null;
-  }
-
-  /// Returns the logout URL
-  static Future<String?> getLogoutUrl() async {
-    if (_logoutUrl == null) {
-      final loaded = await _loadData();
-      if (!loaded) {
-        debugPrint('AuthDataManager => getLogoutUrl() failed');
-        return null;
-      }
-    }
-    assert(_logoutUrl != null);
-    return _logoutUrl;
-  }
-
-  /// Reconstruct the rsaInfo from JSON string
-  static Map<dynamic, dynamic> _getRsaInfo(String rsaJson) {
-    final rsaInfo_ = jsonDecode(rsaJson) as Map<String, dynamic>;
-    final publicKey = rsaInfo_['rsa']['public_key'] as String;
-    final privateKey = rsaInfo_['rsa']['private_key'] as String;
-
-    return {...rsaInfo_, 'rsa': KeyPair(publicKey, privateKey)};
-  }
-
-  /// Retrieve auth data from secure storage
-  static Future<bool> _loadData() async {
-    final dataStr = await secureStorage.read(key: _authDataSecureStorageKey);
-
-    if (dataStr != null) {
-      final dataMap = jsonDecode(dataStr) as Map<String, dynamic>;
-      _logoutUrl = dataMap['logout_url'] as String;
-      _rsaInfo = _getRsaInfo(dataMap['rsa_info'] as String);
-      _authResponse =
-          Credential.fromJson((dataMap['auth_response'] as Map).cast());
-
-      return true;
-    }
-    return false;
-  }
-}
+Future<String?> getWebId() async =>
+    secureStorage.read(key: webIdSecureStorageKey);
 
 /// Generates a list of default folder paths for a given application.
 ///
@@ -548,14 +338,14 @@ class AuthDataManager {
 /// Each string in the list represents a path to a default folder for the application.
 
 Future<List<String>> generateDefaultFolders() async {
-  final appName = await getAppName();
+  final appName = await AppInfo.canonicalName;
   final mainResDir = appName;
 
-  final dataDirLoc = path.join(mainResDir, dataDir);
-  final sharingDirLoc = path.join(mainResDir, sharingDir);
-  final sharedDirLoc = path.join(mainResDir, sharedDir);
-  final encDirLoc = path.join(mainResDir, encDir);
-  final logDirLoc = path.join(mainResDir, logsDir);
+  final dataDirLoc = [mainResDir, dataDir].join('/');
+  final sharingDirLoc = [mainResDir, sharingDir].join('/');
+  final sharedDirLoc = [mainResDir, sharedDir].join('/');
+  final encDirLoc = [mainResDir, encDir].join('/');
+  final logDirLoc = [mainResDir, logsDir].join('/');
 
   final folders = [
     mainResDir,
@@ -574,7 +364,7 @@ Future<List<String>> generateDefaultFolders() async {
 /// Each string in the list represents a path to a default folder for the application.
 
 Future<Map<dynamic, dynamic>> generateDefaultFiles() async {
-  final appName = await getAppName();
+  final appName = await AppInfo.canonicalName;
   final mainResDir = appName;
 
   const encKeyFile = 'enc-keys.ttl';
@@ -582,10 +372,10 @@ Future<Map<dynamic, dynamic>> generateDefaultFiles() async {
   const indKeyFile = 'ind-keys.ttl';
   const permLogFile = 'permissions-log.ttl';
 
-  final sharingDirLoc = path.join(mainResDir, sharingDir);
-  final sharedDirLoc = path.join(mainResDir, sharedDir);
-  final encDirLoc = path.join(mainResDir, encDir);
-  final logDirLoc = path.join(mainResDir, logsDir);
+  final sharingDirLoc = [mainResDir, sharingDir].join('/');
+  final sharedDirLoc = [mainResDir, sharedDir].join('/');
+  final encDirLoc = [mainResDir, encDir].join('/');
+  final logDirLoc = [mainResDir, logsDir].join('/');
 
   final files = {
     sharingDirLoc: [