+++ lib/src/widgets/password_input_screen.dart
@@ -33,7 +33,7 @@ import 'package:flutter/material.dart';
 
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:form_builder_validators/form_builder_validators.dart';
-import 'package:solidpod/src/solid/utils.dart' show saveMasterPassword;
+import 'package:solidpod/src/solid/utils/misc.dart' show saveMasterPassword;
 
 /// MasterPasswordInput is a [StatefulWidget] for user to enter
 /// the master password for data encryption.