+++ CHANGELOG.md
@@ -8,6 +8,7 @@ numbers.
 
 ## 0.6 Future Release
 
++ Refactor code in rest_api.dart and utils.dart. [0.5.20]
 + Add a logout popup for user to logging out. [0.5.19]
 + Add a changekeyPopup widget to open a popup window for changing the key. [0.5.18]
 + Remove TTL string generation function which should be app specific [0.5.17]